(/usr/share/games/fortunes/hitchhiker) % "`Eddies,' said Ford, `in the space-time continuum.' `Ah,' nodded Arthur, `is he? Is he?'"
- Arthur failing in his first lesson of galactic physics
in four years.
