--- src/lib/components/ContextMenu.svelte
@@ -1,208 +1,0 @@
-<!--
-Github @dukenmarga, July 2022
-Context Menu is small menu that displayed when user right-click the mouse on browser.
-Think of it as a way to show Refresh option on Microsoft Windows when right-click on desktop.
-Known bug:
-    - If the browser loads the content for the first time, showMenu set to false.
-    Hence, we cannot get menu.h and menu.y dimension, since context menu has not been available at DOM.
-    The first right click will not shown properly when right-click occurs around the edge (bottom part
-    and right part) of the browser.
-
-Inspired from: Context Menu https://svelte.dev/repl/3a33725c3adb4f57b46b597f9dade0c1?version=3.25.0
--->
-
-<script lang="ts">
-	//var _ = require('lodash');
-	import { izip, cycle } from 'itertools';
-	import { engine } from '../engine/engine.ts';
-	import { gameStatePersisted } from '../../../stores.ts';
-	import { gameParams, gameStateTemp } from '../../states.ts';
-	import { G } from '@svgdotjs/svg.js';
-
-	// pos is cursor position when right click occur
-	let pos = { x: 0, y: 0 };
-	// menu is dimension (height and width) of context menu
-	let menu = { h: 0, y: 0 };
-	// browser/window dimension (height and width)
-	let browser = { h: 0, y: 0 };
-	// showMenu is state of context-menu visibility
-	let showMenu = false;
-	// to display some text
-	let content;
-
-	function rightClickContextMenu(e) {
-		showMenu = true;
-		browser = {
-			w: window.innerWidth,
-			h: window.innerHeight
-		};
-		pos = {
-			x: e.clientX,
-			y: e.clientY
-		};
-		// If bottom part of context menu will be displayed
-		// after right-click, then change the position of the
-		// context menu. This position is controlled by `top` and `left`
-		// at inline style.
-		// Instead of context menu is displayed from top left of cursor position
-		// when right-click occur, it will be displayed from bottom left.
-		if (browser.h - pos.y < menu.h) pos.y = pos.y - menu.h;
-		if (browser.w - pos.x < menu.w) pos.x = pos.x - menu.w;
-	}
-	function onPageClick(e) {
-		// To make context menu disappear when
-		// mouse is clicked outside context menu
-		showMenu = false;
-	}
-	function getContextMenuDimension(node) {
-		// This function will get context menu dimension
-		// when navigation is shown => showMenu = true
-		let height = node.offsetHeight;
-		let width = node.offsetWidth;
-		menu = {
-			h: height,
-			w: width
-		};
-	}
-	function difficulty() {
-		//content.textContent = "Hard/Medium/Easy"
-		gameStatePersisted.TICK = cycle(500, 10000, 100000, 0);
-	}
-	function addPlayers() {
-		let currentLocation = gameState.user.currentLocation;
-		let newBot;
-		newBot = engine.ab();
-		gameState.locationUserMap[currentLocation].push(newBot);
-		console.debug(gameState.locationUserMap[currentLocation].length);
-	}
-	function teleport2City() {}
-	function settings() {
-		content.textContent = 'Settings...';
-	}
-	function restart() {
-		content.textContent = 'Restart game...';
-	}
-	let menuItems = [
-		{
-			name: 'difficulty',
-			onClick: difficulty,
-			displayText: 'Difficulty Level',
-			class: 'fa-solid fa-plus'
-		},
-		{
-			name: 'Add more People',
-			onClick: addPlayers,
-			displayText: 'Add People',
-			class: 'fa-solid fa-square'
-		},
-		{
-			name: 'Teleport 2 different City',
-			onClick: teleport2City,
-			displayText: 'Teleport',
-			class: 'fa-solid fa-magnifying-glass'
-		},
-		{
-			name: 'setting',
-			onClick: settings,
-			displayText: 'Settings',
-			class: 'fa-solid fa-print'
-		},
-		{
-			name: 'restart',
-			onClick: restart,
-			displayText: 'Restart',
-			class: 'fa-solid fa-trash-can'
-		}
-	];
-</script>
-
-<svelte:head>
-	<!-- You can change icon sets according to your taste. Change `class` value in `menuItems` above to represent your icons. -->
-	<!-- <link rel="stylesheet" href="/icon/css/mfglabs_iconset.css"> -->
-	<link
-		rel="stylesheet"
-		href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css"
-		integrity="sha512-KfkfwYDsLkIlwQp6LFnl8zNdLGxu9YAA1QvwINks4PhcElQSvqcyVLLD9aMhXd13uQjoXtEKNosOWaZqXgel0g=="
-		crossorigin="anonymous"
-		referrerpolicy="no-referrer"
-	/>
-</svelte:head>
-
-{#if showMenu}
-	<nav use:getContextMenuDimension style="position: absolute; top:{pos.y}px; left:{pos.x}px">
-		<div class="navbar" id="navbar">
-			<ul>
-				{#each menuItems as item}
-					{#if item.name == 'hr'}
-						<hr />
-					{:else}
-						<li>
-							<button on:click={item.onClick}><i class={item.class}></i>{item.displayText}</button>
-						</li>
-					{/if}
-				{/each}
-			</ul>
-		</div>
-	</nav>
-{/if}
-
-<svelte:window on:contextmenu|preventDefault={rightClickContextMenu} on:click={onPageClick} />
-
-<style>
-	* {
-		padding: 0;
-		margin: 0;
-	}
-	.navbar {
-		display: inline-flex;
-		border: 1px #999 solid;
-		width: 170px;
-		background-color: #fff;
-		border-radius: 10px;
-		overflow: hidden;
-		flex-direction: column;
-	}
-	.navbar ul {
-		margin: 6px;
-	}
-	ul li {
-		display: block;
-		list-style-type: none;
-		width: 1fr;
-	}
-	ul li button {
-		font-size: 1rem;
-		color: #222;
-		width: 100%;
-		height: 30px;
-		text-align: left;
-		border: 0px;
-		background-color: #fff;
-	}
-	ul li button:hover {
-		color: #000;
-		text-align: left;
-		border-radius: 5px;
-		background-color: #eee;
-	}
-	ul li button i {
-		padding: 0px 15px 0px 10px;
-	}
-	ul li button i.fa-square {
-		color: #fff;
-	}
-	ul li button:hover > i.fa-square {
-		color: #eee;
-	}
-	ul li button:hover > i.warning {
-		color: crimson;
-	}
-	:global(ul li button.info:hover) {
-		color: navy;
-	}
-	hr {
-		border: none;
-		border-bottom: 1px solid #ccc;
-		margin: 5px 0px;
-	}
-</style>
